requirements.txt
pandas==2.1.0
matplotlib==3.8.0
seaborn==0.12.2
mysql-connector-python==8.1.0